+++ tests/testapp/test_embed.py
@@ -1,10 +1,11 @@
-from django import test
+import pytest
 from django.template import Context, Template, TemplateSyntaxError
 
 from feincms3_cookiecontrol.embedding import embed
 
 
-class ConsciousEmbedTest(test.TestCase):
+@pytest.mark.django_db
+class TestConsciousEmbed:
     def test_wrap_tag(self):
         template = Template(
             """\
@@ -13,14 +14,13 @@ class ConsciousEmbedTest(test.TestCase):
 """
         )
         html = template.render(Context({}))
-        self.assertIn("f3cc-embed", html)
-        self.assertIn(
-            '<template><iframe src="https://example.com/"></iframe></template>',
-            html,
+        assert "f3cc-embed" in html
+        assert (
+            '<template><iframe src="https://example.com/"></iframe></template>' in html
         )
 
     def test_wrap_tag_template_syntax_error(self):
-        with self.assertRaises(TemplateSyntaxError):
+        with pytest.raises(TemplateSyntaxError):
             Template(
                 """\
 {% load feincms3_cookiecontrol %}{% wrap %}{% endwrap %}
@@ -35,36 +35,35 @@ class ConsciousEmbedTest(test.TestCase):
 """
         )
         html = template.render(Context({}))
-        self.assertIn("f3cc-embed", html)
-        self.assertIn(
-            '<template><iframe src="https://example.com/"></iframe></template>',
-            html,
+        assert "f3cc-embed" in html
+        assert (
+            '<template><iframe src="https://example.com/"></iframe></template>' in html
         )
-        self.assertIn(
-            '<button type="button" class="f3cc-button accept">Show me the good stuff!</button>',
-            html,
+        assert (
+            '<button type="button" class="f3cc-button accept">Show me the good stuff!</button>'
+            in html
         )
 
     def test_embed_vimeo_url(self):
         html = embed("https://vimeo.com/455728498")
-        self.assertIn('href="https://vimeo.com/privacy"', html)
-        self.assertIn("455728498", html)
+        assert 'href="https://vimeo.com/privacy"' in html
+        assert "455728498" in html
 
     def test_embed_youtube_url(self):
         html = embed("https://www.youtube.com/watch?v=dQw4w9WgXcQ")
-        self.assertIn('href="https://policies.google.com/privacy"', html)
-        self.assertIn("dQw4w9WgXcQ", html)
+        assert 'href="https://policies.google.com/privacy"' in html
+        assert "dQw4w9WgXcQ" in html
 
     def test_embed_unknown(self):
-        self.assertEqual(embed("https://example.com"), "")
+        assert embed("https://example.com") == ""
 
     def test_embed_tag(self):
         template = Template("{% load feincms3_cookiecontrol %}{% embed url %}")
         html = template.render(Context({}))
-        self.assertEqual(html, "")
+        assert html == ""
 
         html = template.render(
             Context({"url": "https://www.youtube.com/watch?v=dQw4w9WgXcQ"})
         )
-        self.assertIn('href="https://policies.google.com/privacy"', html)
-        self.assertIn("dQw4w9WgXcQ", html)
+        assert 'href="https://policies.google.com/privacy"' in html
+        assert "dQw4w9WgXcQ" in html